+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Created TileWidget (#256)
 - Added ability to connect or disconnect grid lines (#269)
-- Created a canvas drawing helper trait (#278) 
+- Created a canvas drawing helper trait (#278)
+- Modified `GridWidget` to use `CanvasHelper`.
 
 ## 0.4.21